[F] Debug: Warnings Productos, AuthRouter,RecuperarContrasena, Ventas
